+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- Added `PageBlobItem.getContentCrc64()` to expose CRC64 values returned by page operations.
+
 ### Breaking Changes
 
 ### Bugs Fixed